refactor(ke-graphql): colocate tests + #testing alias (canonical standards)
Bring ke-graphql into full compliance with the canonical web code
standards (code/packaging/tsdoc TTL rules).

Test colocation: move every file out of the top-level test/ dir into
co-located src/**/*.test.ts positions (matching ke and cli/pragma, which
have no top-level test/), and relocate the shared TTL fixtures to
src/testing/ as internal test data.

- test/tbox.test.ts            -> src/lib/tbox/buildTBoxSchema.test.ts
- test/plugin.test.ts          -> src/lib/createSchemaPlugin.test.ts
- test/handler.test.ts         -> src/http/createGraphQLHandler.test.ts
- test/execution.test.ts       -> src/lib/execution/incremental.test.ts
- test/pipeline.test.ts        -> src/lib/compiler/pipeline.test.ts
- test/queries.test.ts         -> src/lib/compiler/queries.test.ts
- test/performance-features.ts -> src/lib/compiler/performance-features.test.ts
- test/hardening.test.ts       -> src/lib/compiler/hardening.test.ts
- test/fixtures.ts             -> src/testing/fixtures.ts (+ #testing barrel)

Test imports now follow the established convention: same-domain modules
via relative paths, cross-domain via the # subpath aliases. Fixtures are
reached through a new #testing alias resolving to the source barrel.

src/testing/ is excluded from the build (tsconfig.build.json) and from
coverage; files: ["dist"] keeps it out of the tarball. It is still
type-checked by tsconfig.json (src/**/*.ts). The # cross-domain imports
field (#compiler/#dataloader/#execution/#hardening/#resolver/#tbox) and
customConditions are preserved unchanged.

- vitest include -> ["src/**/*.test.ts"]; coverage thresholds unchanged
- biome includes: drop stale **/test/**
- delete the now-empty test/ dir

Audited against export/shape, export/declaration, function-verb,
function/purity, function/composition, function/location, constants/file,
types/file, api/stability, export/barrel, library/structure, and the
tsdoc rules: no other violations found (the export-shape and purity
passes were already compliant — deterministic hashing matches the
unannotated hashSources precedent).
