+++ packages/runtime/ke-graphql/biome.json
@@ -1,6 +1,6 @@
 {
   "extends": ["@canonical/biome-config"],
   "files": {
-    "includes": ["**/src/**", "**/test/**", "**/demo/**", "**/*.json"]
+    "includes": ["**/src/**", "**/demo/**", "**/*.json"]
   }
 }
